Split phase2 into modules
