+++ phase2/Cargo.toml
@@ -19,7 +19,6 @@ blake2-rfc = "0.2"
 blake2 = "0.6.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-memmap = "0.7"
 num-bigint = "0.2.3"
 num-traits = "0.2.8"
 itertools = "0.8.1"
