+++ phase2/src/bin/contribute.rs
@@ -1,6 +1,5 @@
 extern crate rand;
 extern crate phase2;
-extern crate memmap;
 extern crate num_bigint;
 extern crate num_traits;
 extern crate blake2;
@@ -12,6 +11,8 @@ use itertools::Itertools;
 
 use std::fs::File;
 use std::fs::OpenOptions;
+
+use phase2::parameters::MPCParameters;
 
 fn main() {
     let args: Vec<String> = std::env::args().collect();
@@ -62,7 +63,7 @@ fn main() {
                             .read(true)
                             .open(in_params_filename)
                             .expect("unable to open.");
-    let mut params = phase2::MPCParameters::read(reader, disallow_points_at_infinity, true).expect("unable to read params");
+    let mut params = MPCParameters::read(reader, disallow_points_at_infinity, true).expect("unable to read params");
 
     println!("Contributing to {}...", in_params_filename);
     let hash = params.contribute(&mut rng);
