+++ phase2/src/hash_writer.rs
@@ -1,0 +1,53 @@
+extern crate blake2_rfc;
+
+use std::io;
+use std::io::Write;
+use blake2_rfc::blake2b::Blake2b;
+
+/// Abstraction over a writer which hashes the data being written.
+pub struct HashWriter<W: Write> {
+    writer: W,
+    hasher: Blake2b
+}
+
+impl Clone for HashWriter<io::Sink> {
+    fn clone(&self) -> HashWriter<io::Sink> {
+        HashWriter {
+            writer: io::sink(),
+            hasher: self.hasher.clone()
+        }
+    }
+}
+
+impl<W: Write> HashWriter<W> {
+    /// Construct a new `HashWriter` given an existing `writer` by value.
+    pub fn new(writer: W) -> Self {
+        HashWriter {
+            writer: writer,
+            hasher: Blake2b::new(64)
+        }
+    }
+
+    /// Destroy this writer and return the hash of what was written.
+    pub fn into_hash(self) -> [u8; 64] {
+        let mut tmp = [0u8; 64];
+        tmp.copy_from_slice(self.hasher.finalize().as_ref());
+        tmp
+    }
+}
+
+impl<W: Write> Write for HashWriter<W> {
+    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+        let bytes = self.writer.write(buf)?;
+
+        if bytes > 0 {
+            self.hasher.update(&buf[0..bytes]);
+        }
+
+        Ok(bytes)
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        self.writer.flush()
+    }
+}
